Refactor NoSeedView
